Native: More work on GPU implementation (both D3D12 and Vulkan) Managed: Working on native engine integration and usage for Image loading.

+++ src/Alimer.Graphics/TextureDescriptor.cs
@@ -35,7 +35,7 @@ public record struct TextureDescriptor
         Width = width;
         Height = height;
         DepthOrArrayLayers = depthOrArrayLayers;
-        MipLevelCount = mipLevelCount == 0 ? (uint)ImageDescription.GetMipLevelCount((int)width, (int)height, dimension == TextureDimension.Texture3D ? (int)depthOrArrayLayers : 1) : mipLevelCount;
+        MipLevelCount = mipLevelCount == 0 ? ImageDescription.GetMipLevelCount(width, height, dimension == TextureDimension.Texture3D ? depthOrArrayLayers : 1u) : mipLevelCount;
         SampleCount = sampleCount;
         Usage = usage;
         CpuAccess = access;
